requirements.txt
fastapi
uvicorn[standard]
httpx
pytest
pytest-randomly
pytest-asyncio
python-multipart
openai-whisper
ruff
mypy
pre-commit
